card_bot/requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
Pillow==10.2.0
python-dotenv==1.0.1
aiosqlite==0.20.0
qrcode[pil]==7.4.2
